ml_service/requirements.txt
fastapi
uvicorn

torch
diffusers
transformers
accelerate
safetensors

opencv-python
numpy
pillow

peft